Combine marketing site tests into a single file
